test: update benchmarks
