+++ packages/benchmark/PARSERS.md
@@ -3,7 +3,7 @@
 ## Single File
 
 ```
-ts-estree x 20,086 ops/sec ±1.66% (83 runs sampled)
-ts-parser x 8,736 ops/sec ±2.30% (83 runs sampled)
+ts-estree x 18,114 ops/sec ±14.31% (76 runs sampled)
+ts-parser x 8,969 ops/sec ±2.70% (81 runs sampled)
 Fastest is ts-estree
 ```
